Modified TextAreaTextEventTest.java as to verify the content change of TextComponent sends TextEvent
